requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openapi
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
langchain_core
fastapi[all]
sse_starlette
uvicorn
langserve
mysql-connector-python
SQLAlchemy
huggingface_hub
cassio
datasets
openai 
tiktoken
PyPDF2
gradio